refactor: rename featureComponent to featuredComponent and update import paths
