+++ templates/featuredComponent.ts
